+++ src/renderer/object/sphere.h
@@ -3,9 +3,9 @@
 
 class Sphere : public Object {
 public:
-    Sphere(glm::vec3 position, float radius, std::string objectName, std::shared_ptr<Material> material) : Object(position, objectName, material), m_radius(radius) {}
-    float getRadius() const { return m_radius; }
-    std::optional<Object::CollisionData> collides(Ray& ray, float threshold = 0.0f) const override;
+    Sphere(glm::dvec3 position, double radius, std::string objectName, std::shared_ptr<Material> material) : Object(position, objectName, material), m_radius(radius) {}
+    double getRadius() const { return m_radius; }
+    std::optional<Object::CollisionData> collides(Ray& ray, double threshold = 0.0) const override;
 protected:
-    float m_radius;
+    double m_radius;
 };
