+++ src/renderer/renderer.cpp
@@ -10,6 +10,6 @@ void Renderer::setCameraSettings(CameraSettings settings) {
     m_camera.setCameraBounceLimit(settings.maxBounces);
 }
 
-std::vector<glm::vec3> Renderer::render() {
+std::vector<glm::dvec3> Renderer::render() {
     return m_camera.render(m_world);
 }
